december-2020-leetcoding-challenge/src/main/java/org/redquark/leetcode/challenge/Problem22_BalancedBinaryTree.java
package org.redquark.leetcode.challenge;

/**
 * @author Anirudh Sharma
 * <p>
 * Given a binary tree, determine if it is height-balanced.
 * <p>
 * For this problem, a height-balanced binary tree is defined as:
 * <p>
 * a binary tree in which the left and right subtrees of every node differ in height by
 * no more than 1.
 * <p>
 * Constraints:
 * <p>
 * The number of nodes in the tree is in the range [0, 5000].
 * -10^4 <= Node.val <= 10^4
 */
public class Problem22_BalancedBinaryTree {

    public boolean isBalanced(Node root) {
        if (root == null) {
            return true;
        }
        // Left height
        int left = height(root.left);
        // Right height
        int right = height(root.right);
        // Check if the subtrees are height-balanced
        return Math.abs(left - right) <= 1 && isBalanced(root.left) && isBalanced(root.right);
    }

    private int height(Node node) {
        if (node == null) {
            return 0;
        }
        return 1 + Math.max(height(node.left), height(node.right));
    }

    static class Node {
        final int data;
        Node left;
        Node right;

        Node(int data) {
            this.data = data;
        }
    }
}
